Move hub templates again
