+++ src/main/steps/applicant1/hub-page/joint/jointTemplateSelector.ts
@@ -1,7 +1,7 @@
 import { CaseWithId } from '../../../../app/case/case';
 import { State, YesOrNo } from '../../../../app/case/definition';
+import { HubTemplate } from '../../../common/hubTemplates';
 import { StateSequence } from '../../../state-sequence';
-import { HubTemplate } from '../hubTemplates';
 
 export const getJointHubTemplate = (
   displayState: StateSequence,
